Move code to modal.app_utils
